Add sidemenu items, update icons, add dropdown
