+++ src/app/app.component.html
@@ -1,5 +1,5 @@
 <ion-app>
-  <app-sidemenu></app-sidemenu>
+  <app-sidemenu #sidemenuRef></app-sidemenu>
   <app-delegated-acc-message *ngIf="isSwitchedToDelegator"></app-delegated-acc-message>
   <ion-router-outlet id="main-content" [ngClass]="{ 'main-outlet': isSwitchedToDelegator }"></ion-router-outlet>
   <app-fy-connection></app-fy-connection>
