+++ src/app/app.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit, EventEmitter, NgZone } from '@angular/core';
+import { Component, OnInit, EventEmitter, NgZone, ElementRef, ViewChild } from '@angular/core';
 import { Platform, MenuController, AlertController, NavController } from '@ionic/angular';
 import { SplashScreen } from '@ionic-native/splash-screen/ngx';
 import { StatusBar } from '@ionic-native/status-bar/ngx';
@@ -28,6 +28,7 @@ import { TrackingService } from './core/services/tracking.service';
 import { LoginInfoService } from './core/services/login-info.service';
 import { PopupService } from './core/services/popup.service';
 import { OrgUserSettings } from './core/models/org_user_settings.model';
+import { SidemenuComponent } from './shared/components/sidemenu/sidemenu.component';
 
 const { App } = Plugins;
 const CapStatusBar = Plugins.StatusBar;
@@ -38,6 +39,8 @@ const CapStatusBar = Plugins.StatusBar;
   styleUrls: ['app.component.scss'],
 })
 export class AppComponent implements OnInit {
+  @ViewChild('sidemenuRef') sidemenuRef: SidemenuComponent;
+
   eou$: Observable<any>;
 
   activeOrg: any;
@@ -158,19 +161,19 @@ export class AppComponent implements OnInit {
     });
   }
 
-  openHelp() {
-    this.menuController.close(); // Closing menu button before opening FreshDesk chat
-    this.freshChatService.openLiveChatSupport();
-  }
+  // openHelp() {
+  //   this.menuController.close(); // Closing menu button before opening FreshDesk chat
+  //   this.freshChatService.openLiveChatSupport();
+  // }
 
-  redirect(route) {
-    this.menuController.close();
-    if (route.indexOf('switch_org') > -1) {
-      this.userEventService.clearCache();
-      globalCacheBusterNotifier.next();
-    }
-    this.router.navigate(route);
-  }
+  // redirect(route) {
+  //   this.menuController.close();
+  //   if (route.indexOf('switch_org') > -1) {
+  //     this.userEventService.clearCache();
+  //     globalCacheBusterNotifier.next();
+  //   }
+  //   this.router.navigate(route);
+  // }
 
   checkAppSupportedVersion() {
     this.deviceService
@@ -201,410 +204,410 @@ export class AppComponent implements OnInit {
       });
   }
 
-  async showSideMenu() {
-    const isLoggedIn = await this.routerAuthService.isLoggedIn();
-    if (!isLoggedIn) {
-      return 0;
-    }
-    const orgs$ = this.offlineService.getOrgs();
-    const currentOrg$ = this.offlineService.getCurrentOrg();
-    const orgSettings$ = this.offlineService.getOrgSettings().pipe(shareReplay(1));
-    const orgUserSettings$ = this.offlineService.getOrgUserSettings();
-    const delegatedAccounts$ = this.offlineService
-      .getDelegatedAccounts()
-      .pipe(map((res) => this.orgUserService.excludeByStatus(res, 'DISABLED')));
-    const deviceInfo$ = this.deviceService.getDeviceInfo();
-    const isSwitchedToDelegator$ = from(this.orgUserService.isSwitchedToDelegator());
+  // async showSideMenu() {
+  //   const isLoggedIn = await this.routerAuthService.isLoggedIn();
+  //   if (!isLoggedIn) {
+  //     return 0;
+  //   }
+  //   const orgs$ = this.offlineService.getOrgs();
+  //   const currentOrg$ = this.offlineService.getCurrentOrg();
+  //   const orgSettings$ = this.offlineService.getOrgSettings().pipe(shareReplay(1));
+  //   const orgUserSettings$ = this.offlineService.getOrgUserSettings();
+  //   const delegatedAccounts$ = this.offlineService
+  //     .getDelegatedAccounts()
+  //     .pipe(map((res) => this.orgUserService.excludeByStatus(res, 'DISABLED')));
+  //   const deviceInfo$ = this.deviceService.getDeviceInfo();
+  //   const isSwitchedToDelegator$ = from(this.orgUserService.isSwitchedToDelegator());
 
-    this.allowedActions$ = orgSettings$.pipe(
-      switchMap((orgSettings) => {
-        const allowedReportsActions$ = this.offlineService.getReportActions(orgSettings);
-        const allowedAdvancesActions$ = this.permissionsService.allowedActions(
-          'advances',
-          ['approve', 'create', 'delete'],
-          orgSettings
-        );
-        const allowedTripsActions$ = this.permissionsService.allowedActions(
-          'trips',
-          ['approve', 'create', 'edit', 'cancel'],
-          orgSettings
-        );
+  //   this.allowedActions$ = orgSettings$.pipe(
+  //     switchMap((orgSettings) => {
+  //       const allowedReportsActions$ = this.offlineService.getReportActions(orgSettings);
+  //       const allowedAdvancesActions$ = this.permissionsService.allowedActions(
+  //         'advances',
+  //         ['approve', 'create', 'delete'],
+  //         orgSettings
+  //       );
+  //       const allowedTripsActions$ = this.permissionsService.allowedActions(
+  //         'trips',
+  //         ['approve', 'create', 'edit', 'cancel'],
+  //         orgSettings
+  //       );
 
-        return forkJoin({
-          allowedReportsActions: allowedReportsActions$,
-          allowedAdvancesActions: iif(
-            () => orgSettings.advance_requests.enabled || orgSettings.advances.enabled,
-            allowedAdvancesActions$,
-            of(null)
-          ),
-          allowedTripsActions: iif(() => orgSettings.trip_requests.enabled, allowedTripsActions$, of(null)),
-        });
-      })
-    );
+  //       return forkJoin({
+  //         allowedReportsActions: allowedReportsActions$,
+  //         allowedAdvancesActions: iif(
+  //           () => orgSettings.advance_requests.enabled || orgSettings.advances.enabled,
+  //           allowedAdvancesActions$,
+  //           of(null)
+  //         ),
+  //         allowedTripsActions: iif(() => orgSettings.trip_requests.enabled, allowedTripsActions$, of(null)),
+  //       });
+  //     })
+  //   );
 
-    this.isConnected$
-      .pipe(
-        switchMap((isConnected) =>
-          forkJoin({
-            orgs: orgs$,
-            currentOrg: currentOrg$,
-            orgSettings: orgSettings$,
-            orgUserSettings: orgUserSettings$,
-            delegatedAccounts: delegatedAccounts$,
-            allowedActions: this.allowedActions$,
-            deviceInfo: deviceInfo$,
-            isSwitchedToDelegator: isSwitchedToDelegator$,
-            isConnected: of(isConnected),
-            eou: this.offlineService.getCurrentUser(),
-          })
-        )
-      )
-      .subscribe((res) => {
-        const orgs = res.orgs;
-        this.activeOrg = res.currentOrg;
-        const orgSettings = res.orgSettings;
-        const orgUserSettings = res.orgUserSettings;
-        const isDelegatee = res.delegatedAccounts.length > 0;
-        this.appVersion = (res.deviceInfo && res.deviceInfo.appVersion) || '1.2.3';
-        const allowedReportsActions = res.allowedActions && res.allowedActions.allowedReportsActions;
-        const allowedAdvancesActions = res.allowedActions && res.allowedActions.allowedAdvancesActions;
-        const allowedTripsActions = res.allowedActions && res.allowedActions.allowedTripsActions;
-        this.isSwitchedToDelegator = res.isSwitchedToDelegator;
-        const isConnected = res.isConnected;
-        this.eou = res.eou;
+  //   this.isConnected$
+  //     .pipe(
+  //       switchMap((isConnected) =>
+  //         forkJoin({
+  //           orgs: orgs$,
+  //           currentOrg: currentOrg$,
+  //           orgSettings: orgSettings$,
+  //           orgUserSettings: orgUserSettings$,
+  //           delegatedAccounts: delegatedAccounts$,
+  //           allowedActions: this.allowedActions$,
+  //           deviceInfo: deviceInfo$,
+  //           isSwitchedToDelegator: isSwitchedToDelegator$,
+  //           isConnected: of(isConnected),
+  //           eou: this.offlineService.getCurrentUser(),
+  //         })
+  //       )
+  //     )
+  //     .subscribe((res) => {
+  //       const orgs = res.orgs;
+  //       this.activeOrg = res.currentOrg;
+  //       const orgSettings = res.orgSettings;
+  //       const orgUserSettings = res.orgUserSettings;
+  //       const isDelegatee = res.delegatedAccounts.length > 0;
+  //       this.appVersion = (res.deviceInfo && res.deviceInfo.appVersion) || '1.2.3';
+  //       const allowedReportsActions = res.allowedActions && res.allowedActions.allowedReportsActions;
+  //       const allowedAdvancesActions = res.allowedActions && res.allowedActions.allowedAdvancesActions;
+  //       const allowedTripsActions = res.allowedActions && res.allowedActions.allowedTripsActions;
+  //       this.isSwitchedToDelegator = res.isSwitchedToDelegator;
+  //       const isConnected = res.isConnected;
+  //       this.eou = res.eou;
 
-        if (res.eou) {
-          Sentry.setUser({
-            id: res.eou.us.email + ' - ' + res.eou.ou.id,
-            email: res.eou.us.email,
-            orgUserId: res.eou.ou.id,
-          });
-        }
+  //       if (res.eou) {
+  //         Sentry.setUser({
+  //           id: res.eou.us.email + ' - ' + res.eou.ou.id,
+  //           email: res.eou.us.email,
+  //           orgUserId: res.eou.ou.id,
+  //         });
+  //       }
 
-        this.freshChatService.setupNetworkWatcher();
+  //       this.freshChatService.setupNetworkWatcher();
 
-        // TODO: remove nested subscribe - mini tech debt
-        this.setupSideMenu(
-          isConnected,
-          orgSettings,
-          orgUserSettings,
-          allowedReportsActions,
-          allowedTripsActions,
-          allowedAdvancesActions,
-          orgs,
-          isDelegatee
-        );
+  //       // TODO: remove nested subscribe - mini tech debt
+  //       this.setupSideMenu(
+  //         isConnected,
+  //         orgSettings,
+  //         orgUserSettings,
+  //         allowedReportsActions,
+  //         allowedTripsActions,
+  //         allowedAdvancesActions,
+  //         orgs,
+  //         isDelegatee
+  //       );
 
-        /* These below conditions have been added to place the divider in the sidenav:-
-        - if 'Advances' is enabled, the divider will be placed under 'Advances',
-        - else if 'Trips' is enabled, the divider will be placed under 'Trips',
-        - else it will be placed under 'Reports'
-      */
-        this.setDividerTitle(orgSettings, orgUserSettings);
-      });
-  }
+  //       /* These below conditions have been added to place the divider in the sidenav:-
+  //       - if 'Advances' is enabled, the divider will be placed under 'Advances',
+  //       - else if 'Trips' is enabled, the divider will be placed under 'Trips',
+  //       - else it will be placed under 'Reports'
+  //     */
+  //       this.setDividerTitle(orgSettings, orgUserSettings);
+  //     });
+  // }
 
-  // TODO: Reduce number of params being passed
-  // eslint-disable-next-line max-params
-  // eslint-disable-next-line max-params-no-constructor/max-params-no-constructor
-  setupSideMenu(
-    isConnected: boolean,
-    orgSettings: any,
-    orgUserSettings: OrgUserSettings,
-    allowedReportsActions: any,
-    allowedTripsActions: any,
-    allowedAdvancesActions: any,
-    orgs,
-    isDelegatee: boolean
-  ) {
-    if (isConnected) {
-      this.setSideMenuOnline(
-        orgSettings,
-        orgUserSettings,
-        allowedReportsActions,
-        allowedTripsActions,
-        allowedAdvancesActions,
-        orgs,
-        isDelegatee
-      );
-    } else {
-      this.setSideMenuOffline(
-        orgSettings,
-        orgUserSettings,
-        allowedReportsActions,
-        allowedTripsActions,
-        allowedAdvancesActions,
-        orgs,
-        isDelegatee
-      );
-    }
-  }
+  // // TODO: Reduce number of params being passed
+  // // eslint-disable-next-line max-params
+  // // eslint-disable-next-line max-params-no-constructor/max-params-no-constructor
+  // setupSideMenu(
+  //   isConnected: boolean,
+  //   orgSettings: any,
+  //   orgUserSettings: OrgUserSettings,
+  //   allowedReportsActions: any,
+  //   allowedTripsActions: any,
+  //   allowedAdvancesActions: any,
+  //   orgs,
+  //   isDelegatee: boolean
+  // ) {
+  //   if (isConnected) {
+  //     this.setSideMenuOnline(
+  //       orgSettings,
+  //       orgUserSettings,
+  //       allowedReportsActions,
+  //       allowedTripsActions,
+  //       allowedAdvancesActions,
+  //       orgs,
+  //       isDelegatee
+  //     );
+  //   } else {
+  //     this.setSideMenuOffline(
+  //       orgSettings,
+  //       orgUserSettings,
+  //       allowedReportsActions,
+  //       allowedTripsActions,
+  //       allowedAdvancesActions,
+  //       orgs,
+  //       isDelegatee
+  //     );
+  //   }
+  // }
 
-  // TODO: Breakdown to make this easier to
-  // eslint-disable-next-line
-  setSideMenuOffline(
-    orgSettings: any,
-    orgUserSettings: OrgUserSettings,
-    allowedReportsActions: any,
-    allowedTripsActions: any,
-    allowedAdvancesActions: any,
-    orgs: any,
-    isDelegatee: boolean
-  ) {
-    this.sideMenuList = [
-      {
-        title: 'Dashboard',
-        isVisible: true,
-        icon: 'fy-dashboard-new',
-        route: ['/', 'enterprise', 'my_dashboard'],
-      },
-      {
-        title: 'Expenses',
-        isVisible: true,
-        icon: 'fy-expenses-new',
-        route: ['/', 'enterprise', 'my_expenses'],
-      },
-      {
-        title: 'Cards',
-        isVisible: orgSettings.corporate_credit_card_settings.enabled,
-        icon: 'fy-cards-new',
-        route: ['/', 'enterprise', 'corporate_card_expenses'],
-        disabled: true,
-      },
-      {
-        title: 'Reports',
-        isVisible: true,
-        icon: 'fy-reports-new',
-        route: ['/', 'enterprise', 'my_reports'],
-        disabled: true,
-      },
-      {
-        title: 'Trips',
-        // eslint-disable-next-line max-len
-        isVisible:
-          orgSettings.trip_requests.enabled &&
-          (!orgSettings.trip_requests.enable_for_certain_employee ||
-            (orgSettings.trip_requests.enable_for_certain_employee &&
-              orgUserSettings.trip_request_org_user_settings.enabled)),
-        icon: 'fy-trips-new',
-        route: ['/', 'enterprise', 'my_trips'],
-        disabled: true,
-      },
-      {
-        title: 'Advances',
-        isVisible: orgSettings.advances.enabled || orgSettings.advance_requests.enabled,
-        icon: 'fy-advances-new',
-        route: ['/', 'enterprise', 'my_advances'],
-        disabled: true,
-      },
-      {
-        title: 'Team Reports',
-        isVisible: allowedReportsActions && allowedReportsActions.approve,
-        icon: 'fy-team-reports-new',
-        route: ['/', 'enterprise', 'team_reports'],
-        cssClass: 'team-trips',
-        disabled: true,
-      },
-      {
-        title: 'Team Trips',
-        isVisible: orgSettings.trip_requests.enabled && allowedTripsActions && allowedReportsActions.approve,
-        icon: 'fy-team-trips-new',
-        route: ['/', 'enterprise', 'team_trips'],
-        disabled: true,
-      },
-      {
-        title: 'Team Advances',
-        isVisible: allowedAdvancesActions && allowedAdvancesActions.approve,
-        icon: 'fy-team-advances-new',
-        route: ['/', 'enterprise', 'team_advance'],
-        disabled: true,
-      },
-    ];
-    this.sideMenuSecondaryList = [
-      {
-        title: 'Live Chat',
-        isVisible:
-          orgUserSettings &&
-          orgUserSettings.in_app_chat_settings &&
-          orgUserSettings.in_app_chat_settings.allowed &&
-          orgUserSettings.in_app_chat_settings.enabled,
-        icon: 'fy-chat',
-        openHelp: true,
-        disabled: true,
-      },
-      {
-        title: 'Settings',
-        isVisible: true,
-        icon: 'fy-settings',
-        route: ['/', 'enterprise', 'my_profile'],
-      },
-      {
-        title: 'Switch Organization',
-        isVisible: orgs.length > 1,
-        icon: 'fy-switch-new',
-        route: ['/', 'auth', 'switch_org', { choose: true }],
-        disabled: true,
-      },
-      {
-        title: 'Delegated Accounts',
-        isVisible: isDelegatee && !this.isSwitchedToDelegator,
-        icon: 'fy-delegate-switch',
-        route: ['/', 'enterprise', 'delegated_accounts'],
-        disabled: true,
-      },
-      {
-        title: 'Switch back to my account',
-        isVisible: this.isSwitchedToDelegator,
-        icon: 'fy-switch',
-        route: ['/', 'enterprise', 'delegated_accounts', { switchToOwn: true }],
-        disabled: true,
-      },
-      {
-        title: 'Help',
-        isVisible: true,
-        icon: 'fy-help-new',
-        route: ['/', 'enterprise', 'help'],
-        disabled: true,
-      },
-    ];
-  }
+  // // TODO: Breakdown to make this easier to
+  // // eslint-disable-next-line
+  // setSideMenuOffline(
+  //   orgSettings: any,
+  //   orgUserSettings: OrgUserSettings,
+  //   allowedReportsActions: any,
+  //   allowedTripsActions: any,
+  //   allowedAdvancesActions: any,
+  //   orgs: any,
+  //   isDelegatee: boolean
+  // ) {
+  //   this.sideMenuList = [
+  //     {
+  //       title: 'Dashboard',
+  //       isVisible: true,
+  //       icon: 'fy-dashboard-new',
+  //       route: ['/', 'enterprise', 'my_dashboard'],
+  //     },
+  //     {
+  //       title: 'Expenses',
+  //       isVisible: true,
+  //       icon: 'fy-expenses-new',
+  //       route: ['/', 'enterprise', 'my_expenses'],
+  //     },
+  //     {
+  //       title: 'Cards',
+  //       isVisible: orgSettings.corporate_credit_card_settings.enabled,
+  //       icon: 'fy-cards-new',
+  //       route: ['/', 'enterprise', 'corporate_card_expenses'],
+  //       disabled: true,
+  //     },
+  //     {
+  //       title: 'Reports',
+  //       isVisible: true,
+  //       icon: 'fy-reports-new',
+  //       route: ['/', 'enterprise', 'my_reports'],
+  //       disabled: true,
+  //     },
+  //     {
+  //       title: 'Trips',
+  //       // eslint-disable-next-line max-len
+  //       isVisible:
+  //         orgSettings.trip_requests.enabled &&
+  //         (!orgSettings.trip_requests.enable_for_certain_employee ||
+  //           (orgSettings.trip_requests.enable_for_certain_employee &&
+  //             orgUserSettings.trip_request_org_user_settings.enabled)),
+  //       icon: 'fy-trips-new',
+  //       route: ['/', 'enterprise', 'my_trips'],
+  //       disabled: true,
+  //     },
+  //     {
+  //       title: 'Advances',
+  //       isVisible: orgSettings.advances.enabled || orgSettings.advance_requests.enabled,
+  //       icon: 'fy-advances-new',
+  //       route: ['/', 'enterprise', 'my_advances'],
+  //       disabled: true,
+  //     },
+  //     {
+  //       title: 'Team Reports',
+  //       isVisible: allowedReportsActions && allowedReportsActions.approve,
+  //       icon: 'fy-team-reports-new',
+  //       route: ['/', 'enterprise', 'team_reports'],
+  //       cssClass: 'team-trips',
+  //       disabled: true,
+  //     },
+  //     {
+  //       title: 'Team Trips',
+  //       isVisible: orgSettings.trip_requests.enabled && allowedTripsActions && allowedReportsActions.approve,
+  //       icon: 'fy-team-trips-new',
+  //       route: ['/', 'enterprise', 'team_trips'],
+  //       disabled: true,
+  //     },
+  //     {
+  //       title: 'Team Advances',
+  //       isVisible: allowedAdvancesActions && allowedAdvancesActions.approve,
+  //       icon: 'fy-team-advances-new',
+  //       route: ['/', 'enterprise', 'team_advance'],
+  //       disabled: true,
+  //     },
+  //   ];
+  //   this.sideMenuSecondaryList = [
+  //     {
+  //       title: 'Live Chat',
+  //       isVisible:
+  //         orgUserSettings &&
+  //         orgUserSettings.in_app_chat_settings &&
+  //         orgUserSettings.in_app_chat_settings.allowed &&
+  //         orgUserSettings.in_app_chat_settings.enabled,
+  //       icon: 'fy-chat',
+  //       openHelp: true,
+  //       disabled: true,
+  //     },
+  //     {
+  //       title: 'Settings',
+  //       isVisible: true,
+  //       icon: 'fy-settings',
+  //       route: ['/', 'enterprise', 'my_profile'],
+  //     },
+  //     {
+  //       title: 'Switch Organization',
+  //       isVisible: orgs.length > 1,
+  //       icon: 'fy-switch-new',
+  //       route: ['/', 'auth', 'switch_org', { choose: true }],
+  //       disabled: true,
+  //     },
+  //     {
+  //       title: 'Delegated Accounts',
+  //       isVisible: isDelegatee && !this.isSwitchedToDelegator,
+  //       icon: 'fy-delegate-switch',
+  //       route: ['/', 'enterprise', 'delegated_accounts'],
+  //       disabled: true,
+  //     },
+  //     {
+  //       title: 'Switch back to my account',
+  //       isVisible: this.isSwitchedToDelegator,
+  //       icon: 'fy-switch',
+  //       route: ['/', 'enterprise', 'delegated_accounts', { switchToOwn: true }],
+  //       disabled: true,
+  //     },
+  //     {
+  //       title: 'Help',
+  //       isVisible: true,
+  //       icon: 'fy-help-new',
+  //       route: ['/', 'enterprise', 'help'],
+  //       disabled: true,
+  //     },
+  //   ];
+  // }
 
-  // eslint-disable-next-line
-  setSideMenuOnline(
-    orgSettings: any,
-    orgUserSettings: OrgUserSettings,
-    allowedReportsActions: any,
-    allowedTripsActions: any,
-    allowedAdvancesActions: any,
-    orgs: any,
-    isDelegatee: boolean
-  ) {
-    this.sideMenuList = [
-      {
-        title: 'Dashboard',
-        isVisible: true,
-        icon: 'fy-dashboard-new',
-        route: ['/', 'enterprise', 'my_dashboard'],
-      },
-      {
-        title: 'Expenses',
-        isVisible: true,
-        icon: 'fy-expenses-new',
-        route: ['/', 'enterprise', 'my_expenses'],
-      },
-      {
-        title: 'Cards',
-        isVisible: orgSettings.corporate_credit_card_settings.enabled,
-        icon: 'fy-cards-new',
-        route: ['/', 'enterprise', 'corporate_card_expenses'],
-      },
-      {
-        title: 'Personal Cards',
-        isVisible: environment.ROOT_URL === 'https://staging.fyle.tech',
-        icon: 'fy-cards-new',
-        route: ['/', 'enterprise', 'personal_cards'],
-      },
-      {
-        title: 'Reports',
-        isVisible: true,
-        icon: 'fy-reports-new',
-        route: ['/', 'enterprise', 'my_reports'],
-      },
-      {
-        title: 'Trips',
-        // eslint-disable-next-line max-len
-        isVisible:
-          orgSettings.trip_requests.enabled &&
-          (!orgSettings.trip_requests.enable_for_certain_employee ||
-            (orgSettings.trip_requests.enable_for_certain_employee &&
-              orgUserSettings.trip_request_org_user_settings.enabled)),
-        icon: 'fy-trips-new',
-        route: ['/', 'enterprise', 'my_trips'],
-      },
-      {
-        title: 'Advances',
-        isVisible: orgSettings.advances.enabled || orgSettings.advance_requests.enabled,
-        icon: 'fy-advances-new',
-        route: ['/', 'enterprise', 'my_advances'],
-      },
-      {
-        title: 'Team Reports',
-        isVisible: allowedReportsActions && allowedReportsActions.approve,
-        icon: 'fy-team-reports-new',
-        route: ['/', 'enterprise', 'team_reports'],
-        cssClass: 'team-trips',
-      },
-      {
-        title: 'Team Trips',
-        isVisible: orgSettings.trip_requests.enabled && allowedTripsActions && allowedReportsActions.approve,
-        icon: 'fy-team-trips-new',
-        route: ['/', 'enterprise', 'team_trips'],
-      },
-      {
-        title: 'Team Advances',
-        isVisible: allowedAdvancesActions && allowedAdvancesActions.approve,
-        icon: 'fy-team-advances-new',
-        route: ['/', 'enterprise', 'team_advance'],
-      },
-    ];
-    this.sideMenuSecondaryList = [
-      {
-        title: 'Live Chat',
-        isVisible:
-          orgUserSettings &&
-          orgUserSettings.in_app_chat_settings &&
-          orgUserSettings.in_app_chat_settings.allowed &&
-          orgUserSettings.in_app_chat_settings.enabled,
-        icon: 'fy-chat',
-        openHelp: true,
-      },
-      {
-        title: 'Settings',
-        isVisible: true,
-        icon: 'fy-settings',
-        route: ['/', 'enterprise', 'my_profile'],
-      },
-      {
-        title: 'Switch Organization',
-        isVisible: orgs.length > 1 && !this.isSwitchedToDelegator,
-        icon: 'fy-switch-new',
-        route: ['/', 'auth', 'switch_org', { choose: true }],
-      },
-      {
-        title: 'Delegated Accounts',
-        isVisible: isDelegatee && !this.isSwitchedToDelegator,
-        icon: 'fy-delegate-switch',
-        route: ['/', 'enterprise', 'delegated_accounts'],
-      },
-      {
-        title: 'Switch back to my account',
-        isVisible: this.isSwitchedToDelegator,
-        icon: 'fy-switch',
-        route: ['/', 'enterprise', 'delegated_accounts', { switchToOwn: true }],
-      },
-      {
-        title: 'Help',
-        isVisible: true,
-        icon: 'fy-help-new',
-        route: ['/', 'enterprise', 'help'],
-      },
-    ];
-  }
+  // // eslint-disable-next-line
+  // setSideMenuOnline(
+  //   orgSettings: any,
+  //   orgUserSettings: OrgUserSettings,
+  //   allowedReportsActions: any,
+  //   allowedTripsActions: any,
+  //   allowedAdvancesActions: any,
+  //   orgs: any,
+  //   isDelegatee: boolean
+  // ) {
+  //   this.sideMenuList = [
+  //     {
+  //       title: 'Dashboard',
+  //       isVisible: true,
+  //       icon: 'fy-dashboard-new',
+  //       route: ['/', 'enterprise', 'my_dashboard'],
+  //     },
+  //     {
+  //       title: 'Expenses',
+  //       isVisible: true,
+  //       icon: 'fy-expenses-new',
+  //       route: ['/', 'enterprise', 'my_expenses'],
+  //     },
+  //     {
+  //       title: 'Cards',
+  //       isVisible: orgSettings.corporate_credit_card_settings.enabled,
+  //       icon: 'fy-cards-new',
+  //       route: ['/', 'enterprise', 'corporate_card_expenses'],
+  //     },
+  //     {
+  //       title: 'Personal Cards',
+  //       isVisible: environment.ROOT_URL === 'https://staging.fyle.tech',
+  //       icon: 'fy-cards-new',
+  //       route: ['/', 'enterprise', 'personal_cards'],
+  //     },
+  //     {
+  //       title: 'Reports',
+  //       isVisible: true,
+  //       icon: 'fy-reports-new',
+  //       route: ['/', 'enterprise', 'my_reports'],
+  //     },
+  //     {
+  //       title: 'Trips',
+  //       // eslint-disable-next-line max-len
+  //       isVisible:
+  //         orgSettings.trip_requests.enabled &&
+  //         (!orgSettings.trip_requests.enable_for_certain_employee ||
+  //           (orgSettings.trip_requests.enable_for_certain_employee &&
+  //             orgUserSettings.trip_request_org_user_settings.enabled)),
+  //       icon: 'fy-trips-new',
+  //       route: ['/', 'enterprise', 'my_trips'],
+  //     },
+  //     {
+  //       title: 'Advances',
+  //       isVisible: orgSettings.advances.enabled || orgSettings.advance_requests.enabled,
+  //       icon: 'fy-advances-new',
+  //       route: ['/', 'enterprise', 'my_advances'],
+  //     },
+  //     {
+  //       title: 'Team Reports',
+  //       isVisible: allowedReportsActions && allowedReportsActions.approve,
+  //       icon: 'fy-team-reports-new',
+  //       route: ['/', 'enterprise', 'team_reports'],
+  //       cssClass: 'team-trips',
+  //     },
+  //     {
+  //       title: 'Team Trips',
+  //       isVisible: orgSettings.trip_requests.enabled && allowedTripsActions && allowedReportsActions.approve,
+  //       icon: 'fy-team-trips-new',
+  //       route: ['/', 'enterprise', 'team_trips'],
+  //     },
+  //     {
+  //       title: 'Team Advances',
+  //       isVisible: allowedAdvancesActions && allowedAdvancesActions.approve,
+  //       icon: 'fy-team-advances-new',
+  //       route: ['/', 'enterprise', 'team_advance'],
+  //     },
+  //   ];
+  //   this.sideMenuSecondaryList = [
+  //     {
+  //       title: 'Live Chat',
+  //       isVisible:
+  //         orgUserSettings &&
+  //         orgUserSettings.in_app_chat_settings &&
+  //         orgUserSettings.in_app_chat_settings.allowed &&
+  //         orgUserSettings.in_app_chat_settings.enabled,
+  //       icon: 'fy-chat',
+  //       openHelp: true,
+  //     },
+  //     {
+  //       title: 'Settings',
+  //       isVisible: true,
+  //       icon: 'fy-settings',
+  //       route: ['/', 'enterprise', 'my_profile'],
+  //     },
+  //     {
+  //       title: 'Switch Organization',
+  //       isVisible: orgs.length > 1 && !this.isSwitchedToDelegator,
+  //       icon: 'fy-switch-new',
+  //       route: ['/', 'auth', 'switch_org', { choose: true }],
+  //     },
+  //     {
+  //       title: 'Delegated Accounts',
+  //       isVisible: isDelegatee && !this.isSwitchedToDelegator,
+  //       icon: 'fy-delegate-switch',
+  //       route: ['/', 'enterprise', 'delegated_accounts'],
+  //     },
+  //     {
+  //       title: 'Switch back to my account',
+  //       isVisible: this.isSwitchedToDelegator,
+  //       icon: 'fy-switch',
+  //       route: ['/', 'enterprise', 'delegated_accounts', { switchToOwn: true }],
+  //     },
+  //     {
+  //       title: 'Help',
+  //       isVisible: true,
+  //       icon: 'fy-help-new',
+  //       route: ['/', 'enterprise', 'help'],
+  //     },
+  //   ];
+  // }
 
-  setDividerTitle(orgSettings: any, orgUserSettings: OrgUserSettings) {
-    this.dividerTitle = 'Reports';
-    if (
-      orgSettings.trip_requests.enabled &&
-      (!orgSettings.trip_requests.enable_for_certain_employee ||
-        (orgSettings.trip_requests.enable_for_certain_employee &&
-          orgUserSettings.trip_request_org_user_settings.enabled))
-    ) {
-      this.dividerTitle = 'Trips';
-    }
-    if (orgSettings.advances.enabled || orgSettings.advance_requests.enabled) {
-      this.dividerTitle = 'Advances';
-    }
-  }
+  // setDividerTitle(orgSettings: any, orgUserSettings: OrgUserSettings) {
+  //   this.dividerTitle = 'Reports';
+  //   if (
+  //     orgSettings.trip_requests.enabled &&
+  //     (!orgSettings.trip_requests.enable_for_certain_employee ||
+  //       (orgSettings.trip_requests.enable_for_certain_employee &&
+  //         orgUserSettings.trip_request_org_user_settings.enabled))
+  //   ) {
+  //     this.dividerTitle = 'Trips';
+  //   }
+  //   if (orgSettings.advances.enabled || orgSettings.advance_requests.enabled) {
+  //     this.dividerTitle = 'Advances';
+  //   }
+  // }
 
   setupNetworkWatcher() {
     const networkWatcherEmitter = new EventEmitter<boolean>();
@@ -631,14 +634,14 @@ export class AppComponent implements OnInit {
     this.checkAppSupportedVersion();
     from(this.routerAuthService.isLoggedIn()).subscribe((loggedInStatus) => {
       if (loggedInStatus) {
-        this.showSideMenu();
+        this.sidemenuRef.showSideMenu();
         this.pushNotificationService.initPush();
       }
     });
 
     this.userEventService.onSetToken(() => {
       setTimeout(() => {
-        this.showSideMenu();
+        this.sidemenuRef.showSideMenu();
       }, 500);
     });
