+++ src/app/fyle/add-edit-advance-request/add-edit-advance-request.page.scss
@@ -15,6 +15,7 @@
       --padding-start: 8px !important;
       --padding-end: 8px !important;
       width: auto !important;
+      color: $black-light;
     }
   }
