+++ src/app/fyle/add-edit-per-diem/add-edit-per-diem.page.scss
@@ -20,6 +20,7 @@
       --padding-start: 8px !important;
       --padding-end: 8px !important;
       width: auto !important;
+      color: $black-light;
     }
   }
