+++ src/app/fyle/my-view-advance-request/my-view-advance-request.page.scss
@@ -1,3 +1,5 @@
+@import '../../../theme/colors.scss';
+
 .view-advance-request {
   &--card {
     padding: 0 16px;
@@ -16,6 +18,7 @@
       --padding-start: 8px !important;
       --padding-end: 8px !important;
       width: auto !important;
+      color: $black-light;
     }
   }
