+++ src/app/fyle/view-per-diem/view-per-diem.page.scss
@@ -26,6 +26,7 @@
       --padding-start: 8px !important;
       --padding-end: 8px !important;
       width: auto !important;
+      color: $black-light;
     }
   }
