+++ src/app/fyle/view-team-advance/view-team-advance.page.scss
@@ -1,3 +1,5 @@
+@import '../../../theme/colors.scss';
+
 .view-advances {
   &--toolbar {
     box-shadow: 0 0 8px rgba(0, 0, 0, 0.1);
@@ -38,6 +40,7 @@
       --padding-start: 8px !important;
       --padding-end: 8px !important;
       width: auto !important;
+      color: $black-light;
     }
   }
