+++ src/app/shared/components/sidemenu/sidemenu-content/sidemenu-content-item/sidemenu-content-item.component.html
@@ -1,0 +1,50 @@
+<div *ngIf="sidemenuItem.title === 'Teams'" class="sidemenu-item__divider"></div>
+<ion-item
+  detail="false"
+  [disabled]="sidemenuItem.disabled"
+  [routerLink]="sidemenuItem.route"
+  [routerLinkActive]="isRoute ? 'sidemenu-item--active' : ''"
+  class="sidemenu-item"
+  [ngClass]="{ 'sidemenu-item--disabled': sidemenuItem.disabled }"
+  (click)="goToRoute(sidemenuItem)"
+>
+  <a class="sidemenu-item__content" [routerLinkActive]="isRoute ? 'sidemenu-item__content--active' : ''">
+    <ion-grid>
+      <ion-row>
+        <ion-col class="sidemenu-item__icon-container">
+          <div>
+            <ion-icon
+              class="sidemenu-item__icon-container__icon"
+              [ngClass]="{
+                'sidemenu-item__icon-container__icon--sm': ['fy-report', 'fy-dashboard-new'].includes(sidemenuItem.icon)
+              }"
+              [src]="'../../../../../assets/svg/' + sidemenuItem.icon + '.svg'"
+              slot="icon-only"
+            ></ion-icon>
+          </div>
+        </ion-col>
+        <ion-col class="sidemenu-item__label-container">
+          <ion-label class="sidemenu-item__label-container__label">{{ sidemenuItem.title }}</ion-label>
+        </ion-col>
+        <ion-col class="sidemenu-item__dropdown-container" *ngIf="sidemenuItem.dropdownOptions?.length">
+          <div>
+            <ion-icon
+              class="sidemenu-item__dropdown-container__icon"
+              [ngClass]="{ 'sidemenu-item__dropdown-container__icon--lg': !sidemenuItem.isDropdownOpen }"
+              [src]="'../../../../../assets/svg/fy-arrow-' + (sidemenuItem.isDropdownOpen ? 'up' : 'down') + '.svg'"
+            ></ion-icon>
+          </div>
+        </ion-col>
+      </ion-row>
+    </ion-grid>
+  </a>
+</ion-item>
+<ng-container *ngIf="sidemenuItem.isDropdownOpen">
+  <ng-container *ngFor="let sidemenuSubItem of sidemenuItem.dropdownOptions">
+    <app-sidemenu-content-item
+      [sidemenuItem]="sidemenuSubItem"
+      (redirect)="goToRoute(sidemenuSubItem)"
+    ></app-sidemenu-content-item>
+  </ng-container>
+</ng-container>
+<div *ngIf="sidemenuItem.title === 'Teams'" class="sidemenu-item__divider"></div>
