+++ src/app/shared/components/sidemenu/sidemenu-content/sidemenu-content.component.ts
@@ -1,4 +1,9 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, Input } from '@angular/core';
+import { MenuController } from '@ionic/angular';
+import { Router } from '@angular/router';
+import { globalCacheBusterNotifier } from 'ts-cacheable';
+import { UserEventService } from 'src/app/core/services/user-event.service';
+import { FreshChatService } from 'src/app/core/services/fresh-chat.service';
 
 @Component({
   selector: 'app-sidemenu-content',
@@ -6,7 +11,33 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./sidemenu-content.component.scss'],
 })
 export class SidemenuContentComponent implements OnInit {
-  constructor() {}
+  @Input() sideMenuList: any[];
+
+  constructor(
+    private router: Router,
+    private userEventService: UserEventService,
+    private menuController: MenuController,
+    private freshChatService: FreshChatService
+  ) {}
 
   ngOnInit(): void {}
+
+  goToRoute(sidemenuItem) {
+    if (sidemenuItem.dropdownOptions?.length) {
+      sidemenuItem.isDropdownOpen = !sidemenuItem.isDropdownOpen;
+      return;
+    }
+    this.menuController.close();
+
+    if (!!sidemenuItem.openLiveChat) {
+      this.freshChatService.openLiveChatSupport();
+      return;
+    }
+
+    if (sidemenuItem.route.indexOf('switch_org') > -1) {
+      this.userEventService.clearCache();
+      globalCacheBusterNotifier.next();
+    }
+    this.router.navigate(sidemenuItem.route);
+  }
 }
