+++ src/app/shared/components/sidemenu/sidemenu.component.html
@@ -1,119 +1,9 @@
 <ion-menu side="start" class="sidemenu" content-id="main-content" [swipeGesture]="false">
-  <ng-container *ngIf="eou$ | async as eou">
-    <ng-container *ngIf="activeOrg$ | async as activeOrg">
-      <app-sidemenu-header [eou]="eou" [activeOrg]="activeOrg"></app-sidemenu-header>
-    </ng-container>
-  </ng-container>
+  <app-sidemenu-header *ngIf="eou && activeOrg" [eou]="eou" [activeOrg]="activeOrg"></app-sidemenu-header>
 
-  <app-sidemenu-content></app-sidemenu-content>
+  <ion-content>
+    <app-sidemenu-content [sideMenuList]="filteredSidemenuList"></app-sidemenu-content>
 
-  <app-sidemenu-footer [appVersion]="appVersion"></app-sidemenu-footer>
+    <app-sidemenu-footer [appVersion]="appVersion"></app-sidemenu-footer>
+  </ion-content>
 </ion-menu>
-
-<!-- <ion-content class="sidemenu--content">
-    <div class="sidemenu--items-wrapper">
-      <ion-list>
-        <div *ngFor="let sideMenuItem of sideMenuList; index as i">
-          <ng-container *ngIf="!sideMenuItem.openHelp">
-            <ion-item
-              detail="false"
-              [disabled]="sideMenuItem.disabled"
-              [routerLink]="sideMenuItem.route"
-              routerLinkActive="item-active"
-              class="sidemenu--items"
-              [class]="sideMenuItem.cssClass"
-              [ngClass]="{ 'sidemenu--items__disabled': sideMenuItem.disabled }"
-              *ngIf="sideMenuItem.isVisible"
-              (click)="redirect(sideMenuItem.route)"
-            >
-              <a class="sidemenu--item" routerLinkActive="sidemenu--active-tab-text">
-                <ion-grid>
-                  <ion-row>
-                    <ion-col size="1">
-                      <div>
-                        <mat-icon class="sidemenu--icon" svgIcon="{{ sideMenuItem.icon }}"></mat-icon>
-                      </div>
-                    </ion-col>
-                    <ion-col size="11">
-                      <ion-label class="sidemenu--label">{{ sideMenuItem.title }}</ion-label>
-                    </ion-col>
-                  </ion-row>
-                </ion-grid>
-              </a>
-            </ion-item>
-            <ng-container *ngIf="allowedActions$ | async as allowedActions">
-              <div
-                *ngIf="
-                  sideMenuItem.title === dividerTitle &&
-                  (allowedActions?.allowedReportsActions?.approve || allowedActions?.allowedAdvancesActions?.approve)
-                "
-                class="sidemenu--divider"
-              ></div>
-            </ng-container>
-          </ng-container>
-        </div>
-      </ion-list> -->
-<!-- SideMenu bottom List start -->
-<!-- <div class="sidemenu--bottom-list">
-        <ion-list>
-          <div *ngFor="let sideMenuSecondaryItem of sideMenuSecondaryList; index as i">
-            <ng-container *ngIf="!sideMenuSecondaryItem.openHelp">
-              <ion-item
-                detail="false"
-                [disabled]="sideMenuSecondaryItem.disabled"
-                [routerLink]="sideMenuSecondaryItem.route"
-                routerLinkActive="item-active"
-                class="sidemenu--items"
-                [class]="sideMenuSecondaryItem.cssClass"
-                [ngClass]="{ 'sidemenu--items__disabled': sideMenuSecondaryItem.disabled }"
-                *ngIf="sideMenuSecondaryItem.isVisible"
-                (click)="redirect(sideMenuSecondaryItem.route)"
-              >
-                <a class="sidemenu--item" routerLinkActive="sidemenu--active-tab-text">
-                  <ion-grid>
-                    <ion-row>
-                      <ion-col size="1">
-                        <div>
-                          <mat-icon class="sidemenu--icon" svgIcon="{{ sideMenuSecondaryItem.icon }}"></mat-icon>
-                        </div>
-                      </ion-col>
-                      <ion-col size="11">
-                        <ion-label class="sidemenu--label">{{ sideMenuSecondaryItem.title }}</ion-label>
-                      </ion-col>
-                    </ion-row>
-                  </ion-grid>
-                </a>
-              </ion-item>
-            </ng-container>
-            <ng-container *ngIf="sideMenuSecondaryItem.openHelp">
-              <ion-item
-                detail="false"
-                [disabled]="sideMenuSecondaryItem.disabled"
-                class="sidemenu--items"
-                [class]="sideMenuSecondaryItem.cssClass"
-                [ngClass]="{ 'sidemenu--items__disabled': sideMenuSecondaryItem.disabled }"
-                *ngIf="sideMenuSecondaryItem.isVisible"
-                (click)="openHelp()"
-              >
-                <a class="sidemenu--item">
-                  <ion-grid>
-                    <ion-row>
-                      <ion-col size="1">
-                        <div>
-                          <mat-icon class="sidemenu--icon" svgIcon="{{ sideMenuSecondaryItem.icon }}"></mat-icon>
-                        </div>
-                      </ion-col>
-                      <ion-col size="11">
-                        <ion-label class="sidemenu--label">{{ sideMenuSecondaryItem.title }}</ion-label>
-                      </ion-col>
-                    </ion-row>
-                  </ion-grid>
-                </a>
-              </ion-item>
-            </ng-container>
-          </div>
-        </ion-list>
-        <div class="sidemenu--appversion" *ngIf="appVersion">v{{ appVersion }}</div>
-      </div>
-    </div>
-  </ion-content> -->
