+++ src/app/shared/components/sidemenu/sidemenu.component.scss
@@ -1,135 +1,6 @@
 @import '../../../../theme/colors.scss';
-$sidenav-bg: #161528;
 
 .sidemenu {
   --ion-background-color: #{$black};
   --ion-text-color: #{$pure-white};
-
-  &--header {
-    // height: 80px !important;
-    position: relative;
-    background-size: 100% 0px !important;
-  }
-
-  &--header-info-wrapper {
-    position: relative;
-    height: 100%;
-    width: 100%;
-    --background: var(--ion-color-fyle-sidenav-bg);
-    border-color: $sidenav-bg;
-  }
-
-  &--header-wrapper {
-    width: 100%;
-    margin: 10px 0;
-    padding: 0 16px;
-    padding-left: 0;
-  }
-
-  &--header-logo {
-    display: block;
-    width: 40px;
-  }
-  &--info-wrapper {
-    margin-left: 10px;
-    width: calc(100% - 60px);
-    color: #fff;
-  }
-  &--user-name {
-    margin: 0;
-    font-size: 16px;
-    line-height: 1;
-    text-transform: uppercase;
-  }
-  &--org-name {
-    margin: 0;
-    margin-top: 5px;
-    color: #a9acbc;
-    font-size: 12px;
-    line-height: 1;
-    text-transform: uppercase;
-  }
-
-  &--content {
-    /*top: 80px !important;*/
-    --background: #161528;
-    padding: 0;
-  }
-
-  &--items-wrapper {
-    display: flex;
-    justify-content: space-between;
-    flex-direction: column;
-    height: 100%;
-  }
-
-  &--divider {
-    margin: 25px 15px;
-    border-top: 1px solid #a9acbc;
-  }
-
-  &--items {
-    --background: #161528;
-    --inner-border-width: 0;
-    --padding-start: 0 !important;
-    --min-height: 50px;
-
-    &.item-active {
-      border-left: 3px solid #ff4d79;
-      --background: #252841;
-    }
-
-    &__disabled {
-      opacity: 0.2;
-    }
-  }
-
-  &--item {
-    width: 100%;
-    border: 0;
-    font-weight: 500;
-    letter-spacing: 0.2px;
-    background: transparent;
-    color: #a9acbc;
-    font-size: 14px;
-    text-decoration: none;
-  }
-
-  &--icon {
-    width: 24px;
-    height: 24px;
-    font-size: 24px;
-    fill: #a9acbc;
-    padding-left: 5px;
-  }
-
-  &--label {
-    padding-left: 15px;
-    padding-top: 3px;
-  }
-
-  &--list-bottom {
-    position: absolute;
-    bottom: 15px;
-    left: 0;
-    width: 100%;
-    background: none !important;
-  }
-
-  &--bottom-list {
-    //margin-top: 50px;
-  }
-
-  &--appversion {
-    padding: 0 10px 10px 0;
-    color: #a9acbc;
-    text-align: right;
-  }
-
-  &--active-tab-text {
-    mat-icon {
-      fill: white !important;
-    }
-    color: white;
-  }
 }
