+++ src/theme/colors.scss
@@ -53,6 +53,7 @@ $green-1: #24a148 !default;
 $blue: #0077ee !default;
 $blue-1: #2896f3 !default;
 $blue-2: #ebf3fc !default;
+$blue-3: #252841 !default;
 $blue-black: #2c304e !default;
 $light-blue: #f7f7fb !default;
 $dark-blue: #1c1f36 !default;
